Changed default value for parameter String &name in classes, derived from GraphParameters, from _T("Untitled") -> s_defaultName (which is declared as static const TCHAR * in PersistentData) More private membervariables, setters and getters Renamed DiffEquationSystemDescription -> DiffEquaionDescriptionArray Modified returntype of ParserTree::decodeErrorString from int->UINT Same modofication for SourcePosition::findCharIndex Added functions addEquation and removeAllEquations to DiffEquationGraphParameters Changed m_currentAdjustSet in CDiffEquationGraphDlg from CompactIntArray -> BitSet Added parameter SHOWXCOORDINATE in call to MoveablePoint::constructor in function GraphZeroesResult::getMoveablePointArray()

+++ Lib/Src/MFCUtil/ExprDialog.cpp
@@ -85,16 +85,16 @@ bool CExprDialog::validateMinMax(int id, double min, double max) {
 
 void CExprDialog::showExprError(const String &msg, int id) {
   try {
-    String    errorMsg  = msg;
-    int       charIndex = ParserTree::decodeErrorString(getExprString(id), errorMsg);
-    const int prefixLen = (int)getCommonExprString().length();
+    String     errorMsg  = msg;
+    UINT       charIndex = ParserTree::decodeErrorString(getExprString(id), errorMsg);
+    const UINT prefixLen = (UINT)getCommonExprString().length();
     if(charIndex < prefixLen) {
       gotoExpr(getCommonExprFieldId());
-      getExprField(getCommonExprFieldId())->SetSel(charIndex, charIndex);
+      getExprField(getCommonExprFieldId())->SetSel((int)charIndex, (int)charIndex);
     } else {
       gotoExpr(id);
       charIndex -= prefixLen;
-      getExprField(id)->SetSel(charIndex, charIndex);
+      getExprField(id)->SetSel((int)charIndex, (int)charIndex);
     }
     showWarning(errorMsg);
   } catch(Exception) { // ignore Exception, and just show msg
